requirements.txt
streamlit==1.46.1
numpy>=1.26.4,<1.27
pandas>=2.3.1,<2.4
plotly>=6.2.0,<6.3
py3Dmol>=2.0.4,<2.1

# Materials Project / crystallography
mp-api>=0.45.8,<0.46
pymatgen>=2025.10.7,<2025.11
monty>=2025.1.9  # align with pymatgen requirements
maggma<=0.72.0
spglib>=2.5  # required by pymatgen>=2025.10.7
hra-api-client>=0.17.0,<0.18

# ML / data science
tensorflow>=2.14,<2.19
scikit-learn>=1.3.1,<1.4
joblib>=1.3.2,<1.4
rdkit-pypi<=2023.3.1b1

# Utilities
python-dotenv>=1.0.1,<1.1
requests>=2.31.0,<2.32
tqdm>=4.67.1,<4.68